+++ CHANGELOG.md
@@ -28,6 +28,7 @@
 * ECS - "Change instance" modal adaptation for central ordering enabled. Refs UIOR-1290.
 * ECS - Align permissions on consortium locations usage. Refs UIOR-1295.
 * Add title token to printed routing list configuration. Refs UIOR-1287.
+* Display total credited in composite orders. Refs UIOR-1282.
 
 ## [6.0.4](https://github.com/folio-org/ui-orders/tree/v6.0.4) (2024-04-25)
 [Full Changelog](https://github.com/folio-org/ui-orders/compare/v6.0.3...v6.0.4)